Fix org member avatar (#102)
* 改进组织成员管理逻辑。

* 改进程序。

* 改进成员代码。

* 改进资源代码

* 改进代码。

* 清除过时的方法。

* 改进。

+++ src/AlphaID/AlphaIdPlatform/Subjects/OrganizationMemberManager.cs
@@ -1,6 +1,5 @@
-using IdSubjects;
 using System.Diagnostics;
-using AlphaIdPlatform.Identity;
+using System.DirectoryServices.ActiveDirectory;
 
 namespace AlphaIdPlatform.Subjects;
 
@@ -11,69 +10,16 @@ namespace AlphaIdPlatform.Subjects;
 /// Init Organization Member Manager via Organization Member store.
 /// </remarks>
 /// <param name="store"></param>
-public class OrganizationMemberManager(IOrganizationMemberStore store)
+/// <param name="organizationManager"></param>
+public class OrganizationMemberManager(IOrganizationMemberStore store, OrganizationManager organizationManager)
 {
-    /// <summary>
-    /// Get the member.
-    /// </summary>
-    /// <param name="person"></param>
-    /// <param name="organization"></param>
-    /// <returns></returns>
-    public Task<OrganizationMember?> GetMemberAsync(string person, string organization)
-    {
-        OrganizationMember? result =
-            store.OrganizationMembers.FirstOrDefault(
-                p => p.PersonId == person && p.OrganizationId == organization);
-        return Task.FromResult(result);
-    }
-
-    /// <summary>
-    /// 获取可见的组织成员。
-    /// </summary>
-    /// <remarks>
-    /// 该方法会考虑组织成员的Visibility设置，根据visitor来决定是否在结果集合中包括这些成员。
-    /// </remarks>
-    /// <param name="organization">AN organization that members to get.</param>
-    /// <param name="visitor">The person who access this system. null if anonymous access.</param>
-    /// <returns></returns>
-    public Task<IEnumerable<OrganizationMember>> GetVisibleMembersAsync(Organization organization,
-        NaturalPerson? visitor)
-    {
-        IQueryable<OrganizationMember>? members =
-            store.OrganizationMembers.Where(p => p.OrganizationId == organization.Id);
-        Debug.Assert(members != null);
-        //假定可见级别为Public。
-        var visibilityLevel = MembershipVisibility.Public;
-        if (visitor != null)
-        {
-            //如果已登录，降为AuthenticatedUser
-            visibilityLevel = MembershipVisibility.AuthenticatedUser;
-            //如果访问者是该组织成员，则降为Private
-            if (members.Any(m => m.PersonId == visitor.Id))
-                visibilityLevel = MembershipVisibility.Private; //Visitor is a member of the organization.
-        }
-        // 过滤出成员可见级别大于等于访问者最低可见级别的成员。
-        return Task.FromResult(members.Where(m => m.Visibility >= visibilityLevel).AsEnumerable());
-    }
-
-    /// <summary>
-    /// Get organization members.
-    /// </summary>
-    /// <param name="organization">Organization</param>
-    /// <returns></returns>
-    public Task<IEnumerable<OrganizationMember>> GetMembersAsync(Organization organization)
-    {
-        IQueryable<OrganizationMember> members =
-            store.OrganizationMembers.Where(p => p.OrganizationId == organization.Id);
-        return Task.FromResult(members.AsEnumerable());
-    }
-
     /// <summary>
     /// 以访问者visitor的视角检索指定用户的组织成员身份。
     /// </summary>
     /// <param name="personId">要检索组织成员身份的目标用户。</param>
     /// <param name="visitorId">访问者。如果传入null，代表匿名访问者。</param>
     /// <returns></returns>
+    [Obsolete]
     public IQueryable<OrganizationMember> GetVisibleMembersOf(string personId, string? visitorId)
     {
         //获取目标person的所有组织身份。
@@ -93,52 +39,49 @@ public class OrganizationMemberManager(IOrganizationMemberStore store)
     }
 
     /// <summary>
-    /// 获取个人的组织成员身份。
+    /// 
     /// </summary>
-    /// <param name="person"></param>
+    /// <param name="organizationId"></param>
+    /// <param name="userId"></param>
+    /// <param name="visibility"></param>
     /// <returns></returns>
-    public Task<IEnumerable<OrganizationMember>> GetMembersOfAsync(NaturalPerson person)
-    {
-        IQueryable<OrganizationMember> members = store.OrganizationMembers.Where(p => p.PersonId == person.Id);
-        return Task.FromResult(members.AsEnumerable());
-    }
-
-    /// <summary>
-    /// </summary>
-    /// <param name="member"></param>
-    /// <returns></returns>
-    public async Task<OrganizationOperationResult> CreateAsync(OrganizationMember member)
+    public async Task<OrganizationMember> Join(string organizationId, string userId, MembershipVisibility visibility)
     {
         if (store.OrganizationMembers.Any(p =>
-                p.OrganizationId == member.OrganizationId && p.PersonId == member.PersonId))
-            return OrganizationOperationResult.Failed(Resources.Membership_exists);
-        return await store.CreateAsync(member);
+                p.OrganizationId == organizationId && p.PersonId == userId))
+            throw new InvalidOperationException(Resources.MembershipExists);
+        var organization = await organizationManager.FindByIdAsync(organizationId) ?? throw new ArgumentException(Resources.OrganizationNotFound, nameof(organizationId));
+        var member = new OrganizationMember(organization, userId, visibility);
+        var result = await store.CreateAsync(member);
+        if (!result.Succeeded)
+            throw new InvalidOperationException(Resources.MembershipCreateFailed);
+        return member;
     }
 
     /// <summary>
-    /// Take person leave out the organization.
+    /// 从指定组织中移除成员。
     /// </summary>
-    /// <param name="member"></param>
-    /// <returns></returns>
-    public async Task<OrganizationOperationResult> LeaveOrganizationAsync(OrganizationMember member)
+    /// <remarks>如果为设置force为true，当用户是组织的唯一所有者时，不能离开组织。这种情况下，操作将失败，并且结果将包含一条错误消息，指示最后一位所有者无法离开。</remarks>
+    /// <param name="organizationId">要从中移除用户的组织唯一标识符。</param>
+    /// <param name="userId">要从组织中移除的用户唯一标识符。</param>
+    /// <param name="force">强制离开，默认为false。如果设为true，即使该成员为组织的最后1个所有者，也将其离开组织。</param>
+    /// <returns>一个 <see cref="OrganizationOperationResult"/>，指示操作结果。如果用户已成功移除或本就不是成员，则返回 <see cref="OrganizationOperationResult.Success"/>。如果操作无法完成，则返回带有相应错误消息的 <see cref="OrganizationOperationResult.Failed"/>。</returns>
+    public async Task<OrganizationOperationResult> Leave(string organizationId, string userId, bool force = false)
     {
-        IQueryable<OrganizationMember> members =
-            store.OrganizationMembers.Where(m => m.OrganizationId == member.OrganizationId);
+        var orgMembers = store.OrganizationMembers.Where(m => m.OrganizationId == organizationId);
+        var member =
+            orgMembers.FirstOrDefault(p => p.PersonId == userId);
+        if (member == null)
+            return OrganizationOperationResult.Failed(Resources.OrganizationMemberNotFound);
 
-        if (member.IsOwner && members.Count(m => m.IsOwner) <= 1)
+        if (force)
+            return await store.DeleteAsync(member);
+
+        // 如果用户是组织所有者，并且该组织只有一个所有者，则不能离开组织。
+        if (member.IsOwner && orgMembers.Count(m => m.OrganizationId == organizationId && m.IsOwner) <= 1)
             return OrganizationOperationResult.Failed(Resources.LastOwnerCannotLeave);
 
         return await store.DeleteAsync(member);
-    }
-
-    /// <summary>
-    /// 移除用户成员身份，无论用户是否是组织所有者也是如此。
-    /// </summary>
-    /// <param name="member"></param>
-    /// <returns></returns>
-    public Task<OrganizationOperationResult> RemoveAsync(OrganizationMember member)
-    {
-        return store.DeleteAsync(member);
     }
 
     /// <summary>
@@ -149,39 +92,5 @@ public class OrganizationMemberManager(IOrganizationMemberStore store)
     public Task<OrganizationOperationResult> UpdateAsync(OrganizationMember member)
     {
         return store.UpdateAsync(member);
-    }
-
-    /// <summary>
-    /// 
-    /// </summary>
-    /// <param name="member"></param>
-    /// <returns></returns>
-    public async Task<OrganizationOperationResult> SetOwner(OrganizationMember member)
-    {
-        var members = await GetMembersAsync(member.Organization);
-        if (members.Count(m => m.IsOwner) <= 5)
-        {
-            member.IsOwner = true;
-            //todo need log to audit log.
-            return await UpdateAsync(member);
-        }
-        return OrganizationOperationResult.Failed(string.Format(Resources.Max_owners_in_the_organization, 5));
-    }
-
-    /// <summary>
-    /// 
-    /// </summary>
-    /// <param name="member"></param>
-    /// <returns></returns>
-    public async Task<OrganizationOperationResult> UnsetOwner(OrganizationMember member)
-    {
-        var members = await GetMembersAsync(member.Organization);
-        if (members.Count(m => m.IsOwner) > 1)
-        {
-            member.IsOwner = false;
-            //todo need log to audit log.
-            return await UpdateAsync(member);
-        }
-        return OrganizationOperationResult.Failed(string.Format(Resources.Max_owners_in_the_organization, 5));
     }
 }
